Clean-up of some test cases.
git-svn-id: svn://svn.code.sf.net/p/ifolder/code/trunk@853 db2431b9-df06-0410-971c-a38ea9a75485

+++ simias/src/Invitation/Tests/InvitationTests.cs
@@ -60,13 +60,11 @@ namespace Simias.Invite.Tests
 		{
 			// the test store
 			string path = Path.GetFullPath(".invitation" + id++);
-
-			// clear any stale store
-			if (Directory.Exists(path))
-			{
-				Directory.Delete(path, true);
-			}
-
+			
+			// clean-up
+			if (Directory.Exists(path)) Directory.Delete(path, true);
+			
+			// configuration
 			Configuration config = new Configuration(path);
 
 			// store
